# Submission Window Management Job