- Creation of Submission Window Job
  - SubmissionReminderJob
  - WindowNotificationJob
  - CloseWindowJob

## Pull Request

```
**Delete these instructions!!!**

Please ensure your pull request meets the following criteria:
- Named "feature/## issue title"
- Team label
- Current sprint milestone
- Assigned to yourself
- Add reviewers if possible
- No estimate
- No Epic
- No Release
- Connected to the issue via ZenHub

```
# Submission Window Management Job

## Overview
This implementation provides a Quartz.NET job for managing submission windows and handling email notifications based on specific event types (SUBMISSIONREMINDER and WINDOWNOTIFICATION) in the ECMPS system.

## Requirements
- .NET 8.0 SDK or later
- Quartz.NET (compatible with .NET 8)
- SQL Server (for Quartz job store)

## Features
- **Event Types**:
  - SUBMISSIONREMINDER (Events 151, 152)
    - Sends to SMPQAEM agents
    - Falls back to DR if no SMPQAEM agents exist
  - WINDOWNOTIFICATION (Events 155, 156)
    - Sends to DR, ADR, and SMPQAEM agents
  - CLOSEWINDOW
    - Closes the specified submission window
    - Sends notifications to DR, ADR, and SMPQAEM agents
    - Prevents further submissions for the closed window
- **Technical Features**:
  - Built for .NET 8
  - Native JSON serialization
  - Raw string literals for email templates
  - Thread pool optimization
  - Dependency Injection support
  - Comprehensive error handling and logging
  - Integration with Silkier Quartz UI
  - Environment variable configuration

## Required Environment Variables

```bash
# Auth API URL
EASEY_AUTH_API=https://((apiHost))/api/auth-mgmt

# Email Configuration
EASEY_QUARTZ_SCHEDULER_EMAIL=((email))
EASEY_QUARTZ_SCHEDULER_SMTP_HOST=((smtpHost))
EASEY_QUARTZ_SCHEDULER_SMTP_PORT=((smtpPort))

# Optional UI Configuration
EASEY_QUARTZ_SCHEDULER_DISPLAY_UI=True  # Enable Silkier Quartz UI

# Other Application Settings
EASEY_QUARTZ_SCHEDULER_TITLE=((title))
EASEY_QUARTZ_SCHEDULER_HOST=((host))
EASEY_QUARTZ_SCHEDULER_ENV=((environment))
```

## Project Structure
```
admin/Jobs/
├── SubmissionWindowManagementJob.cs
├── Extensions/
│   └── SubmissionWindowJobExtensions.cs
└── Tests/
    ├── SubmissionWindowManagementJobTests.cs
    └── README-Testing.md
```

## Setup Instructions

### 1. Dependencies
Required NuGet packages:
```xml
<ItemGroup>
    <PackageReference Include="Quartz" Version="3.8.0" />
    <PackageReference Include="Microsoft.Extensions.Http" Version="8.0.0" />
    <PackageReference Include="Microsoft.Extensions.Logging" Version="8.0.0" />
    <PackageReference Include="Microsoft.Extensions.Configuration" Version="8.0.0" />
</ItemGroup>
```

### 2. Registration
In your `Startup.cs`:
```csharp
using EaseyQuartz.Admin.Extensions;

public void ConfigureServices(IServiceCollection services)
{
    // Add HTTP client for Auth API
    services.AddHttpClient("AuthApi");

    // Register the Quartz job
    services.RegisterWithQuartz();
}
```

## Using the Silkier Quartz UI

1. Enable the UI by setting `EASEY_QUARTZ_SCHEDULER_DISPLAY_UI=True`
2. Access the UI at `/quartz`
3. Authenticate with valid CDX credentials
4. Create and manage job instances:

### Creating Job Instances
1. Navigate to the "Jobs" tab
2. Create instances for:
   - SubmissionReminderJob (Events 151, 152)
   - WindowNotificationJob (Events 155, 156)
   - CloseWindowJob (for closing submission windows)

When creating a CloseWindowJob instance:
1. Set the WindowId in the job data map
2. Configure the trigger based on when the window should close
3. The job will:
   - Close the submission window via API
   - Send notifications to relevant recipients
   - Log the closure in the job history

### Setting Up Triggers
1. Navigate to the "Triggers" tab
2. Create triggers using:
   - Cron expressions for recurring schedules:
     ```
     WeekdayMorning: "0 0 9 ? * MON-FRI"
     WeekdayEvening: "0 0 17 ? * MON-FRI"
     EndOfDay: "0 0 23 ? * *"
     StartOfDay: "0 0 7 ? * *"
     ```
   - Simple triggers for one-time events:
     - Useful for CLOSEWINDOW events at specific dates/times
     - Example: Set exact date/time when a window should close

## .NET 8 Features Used

1. **Performance Optimizations**
   - Optimized thread pool configuration
   - Native JSON serialization
   - Improved HTTP client factory integration

2. **Language Features**
   - Raw string literals for email templates
   - Record types for data models
   - Pattern matching enhancements
   - Required members

3. **Runtime Features**
   - Native AOT compatibility
   - Improved diagnostic capabilities
   - Enhanced logging integration

## Error Handling
- Environment variable validation on startup
- Comprehensive error logging
- Graceful handling of:
  - Recipient API failures
  - Window closure API failures
  - Individual email sending failures
- Failed notifications don't prevent window closure
- Window closure failures are properly logged and reported

## Monitoring
Monitor job execution through:
1. Application logs
2. Silkier Quartz UI history
3. Recent History plugin data

## Support
For issues:
1. Check environment variables are properly set
2. Verify Auth API connectivity
3. Check SMTP server configuration
4. Review application logs for specific error messages
5. Verify Silkier Quartz UI authentication

----
# Testing Strategy for Submission Window Management Job (.NET 8)

## Overview

The testing strategy combines automated unit tests with manual UI testing to ensure comprehensive coverage of all functionality, leveraging .NET 8's enhanced testing capabilities.

## Test Structure

Tests are organized in the `admin/Jobs/Tests` directory, mirroring the production code structure:
```
admin/Jobs/
├── Tests/
│   ├── SubmissionWindowManagementJobTests.cs
│   └── README-Testing.md
├── SubmissionWindowManagementJob.cs
└── Extensions/
    └── SubmissionWindowJobExtensions.cs
```

## Unit Tests

### Test Dependencies (.NET 8)
```xml
<ItemGroup>
    <PackageReference Include="xunit" Version="2.6.x" />
    <PackageReference Include="Moq" Version="4.20.x" />
    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.8.x" />
    <PackageReference Include="coverlet.collector" Version="6.0.x">
      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
      <PrivateAssets>all</PrivateAssets>
    </PackageReference>
    <PackageReference Include="Microsoft.Extensions.Http" Version="8.0.0" />
    <PackageReference Include="Microsoft.Extensions.Configuration" Version="8.0.0" />
</ItemGroup>
```

### Running Tests

```bash
# Run tests with .NET 8 features
dotnet test --framework net8.0

# Run with code coverage
dotnet test --collect:"XPlat Code Coverage"

# Run specific test category
dotnet test --filter "Category=SubmissionWindow"

# Run tests with detailed diagnostics
dotnet test -l "console;verbosity=detailed"
```

### Test Coverage Areas

1. Configuration Tests
   - Environment variable validation
   - .NET 8 configuration binding
   - Null validation using new ArgumentNullException methods

2. Submission Reminder (TT#4475, TT#5826)
   - Recipient filtering rules
   - SMPQAEM agent prioritization
   - DR fallback logic
   - Exclusion rules

3. Window Notification (TT#5826)
   - DR, ADR, SMPQAEM recipient inclusion
   - Exclusion rules
   - Active/inactive agent handling

4. Close Window (TT#6241)
   - Window closure API integration
   - Notification sending
   - Error handling
   - WindowId parameter validation

### .NET 8 Testing Features Used

1. Improved Mocking
   ```csharp
   // Using new HttpClient factory mocking
   services.AddHttpClient("AuthApi").ConfigurePrimaryHttpMessageHandler(() => mockHandler.Object);
   ```

2. Enhanced Assertions
   ```csharp
   // New assertion capabilities
   Assert.Multiple(() =>
   {
       Assert.NotNull(result);
       Assert.Equal(expected.Type, result.Type);
       Assert.True(result.IsActive);
   });
   ```

3. Parallel Test Execution
   ```csharp
   // Collection definition for parallel execution
   [Collection("Submission Window Tests")]
   public class SubmissionWindowManagementJobTests
   {
       // Tests can run in parallel
   }
   ```

## Manual Testing via UI

### 1. Basic Job Setup Testing
- Enable UI: Set `EASEY_QUARTZ_SCHEDULER_DISPLAY_UI=True`
- Access `/quartz` and authenticate
- Create job instances:
  - SubmissionReminderJob
  - WindowNotificationJob
  - CloseWindowJob

### 2. Recipient Integration Testing
Test scenarios for each event type:
```plaintext
SUBMISSIONREMINDER:
├── Active SMPQAEM agents exist
│   └── Only SMPQAEM agents receive notifications
├── No active SMPQAEM agents
│   └── DR receives notifications
└── Mixed recipient types
    └── Verify exclusion rules

WINDOWNOTIFICATION:
├── All recipient types present
│   └── Verify correct filtering
└── Specific scenarios
    ├── Only DR present
    ├── Only ADR present
    └── Only SMPQAEM present

CLOSEWINDOW:
├── Valid window closure
│   ├── API call successful
│   └── Notifications sent
└── Error scenarios
    ├── Invalid WindowId
    ├── Already closed
    └── API failures
```

## Performance Testing

### 1. Load Testing
```bash
# Using k6 for load testing
k6 run load-test.js --vus 10 --duration 30s
```

### 2. Concurrency Testing
- Multiple simultaneous job executions
- Parallel notification sending
- API response handling under load

## Continuous Integration

### GitHub Actions Workflow
```yaml
name: .NET 8 Tests
on: [push, pull_request]
jobs:
  test:
    runs-on: ubuntu-latest
    steps:
      - uses: actions/checkout@v3
      - name: Setup .NET 8
        uses: actions/setup-dotnet@v3
        with:
          dotnet-version: '8.0.x'
      - name: Test
        run: |
          dotnet test --collect:"XPlat Code Coverage"
          dotnet test /p:CollectCoverage=true /p:CoverletOutputFormat=opencover
```

## Testing Notes

### 1. Environment Setup
- Use test SMTP server (like Papercut-SMTP for local testing)
- Mock Auth API responses
- Configure test recipient data

### 2. Test Data Management
- Create test fixtures for recipient configurations
- Maintain test window IDs
- Setup mock API responses

### 3. Error Scenarios
- Network failures
- API timeouts
- Invalid configurations
- Missing parameters

### 4. Performance Considerations
- Monitor memory usage
- Track execution time
- Verify thread pool efficiency
- Test email sending performance

## Coverage Requirements

1. Code Coverage Targets:
   - Overall coverage: ≥80%
   - Critical paths: 100%
   - Business logic: ≥90%

2. Critical Paths:
   - Recipient filtering logic
   - Window closure operations
   - Email notification sending
   - Error handling routines

## Debugging Tips

1. Use .NET 8 diagnostic tools:
```bash
dotnet-trace collect --providers Microsoft.Extensions.Logging
```

2. Enable detailed logging:
```json
{
  "Logging": {
    "LogLevel": {
      "Default": "Debug",
      "EaseyQuartz.Admin.Jobs": "Trace"
    }
  }
}
```

3. Monitor HTTP calls:
```csharp
services.AddHttpClient("AuthApi")
    .AddHttpMessageHandler(() => new HttpLoggingHandler());